add back method duplication to fix fqcn issue
